Removed replicated code
